product/juno: fix MISRA C 2012 rule 10.1 defects
In this patch MISRA C 2012 rule 10.1 defects found
in juno moules  are fixed. This includes
adding cast and changing variables' data types where
necessary.

+++ product/juno/src/juno_id.c
@@ -127,7 +127,7 @@ int juno_id_get_variant(enum juno_idx_variant *variant)
     }
 
     /* NOTE: SYSTOP must be powered ON for this register to be accessible. */
-    var = (V2M_SYS_REGS->ID & V2M_SYS_REGS_ID_VAR_MASK) >>
+    var = (V2M_SYS_REGS->ID & (uint32_t)V2M_SYS_REGS_ID_VAR_MASK) >>
         V2M_SYS_REGS_ID_VAR_POS;
 
     if (!fwk_expect(var < JUNO_IDX_VARIANT_COUNT))
